Implement first UI draft (timetable and timeslots)
Implemented a first draft of the UI. Almost everything was done
using Vuetify.
- Created the TimeTable.vue component which handles the list of
time slots with the lectures. The timetable uses mock data for
now. It is fixed on the right part of the screen and uses the
v-navigation-drawer from Vuetify. It stays clipped on the desktop,
but acts like a normal nav drawer on small devices.
- Created the TimeSlot.vue component which is responsible for a
single time slot. Multiple TimeSlot components make up the time
table.