Create vuex store and implement axios api call
Created a vuex store to have global access to the data on the
lectures. The data is retrieved from the api using axios.
Everything related to the store is contained in the index.js file
for now; in the future it would be better to separate the modules
in different files.