Implement display plants and update timetable
After trying out different alternatives, decided to create a component
for each svg file. This is because I needed to add @click events to
some elements of the svg; this seems to not be possible when loading
the svg from an external file.
Implemented the behaviour for the click event on the building C. It
updates the entries in the timetable (through the store) and displays
the respective svg for the building with its classrooms (at floor 0).
Added breadcrumbs element and a pagination element; none of them works
at the moment.
Finally fixed the positioning and the scaling of the svg file and the
other elements in the content area.