Implement breadcrumb and pagination for floors
Implemented the functionality for the breadcrumb, which keeps track
of the navigation; it uses a stack for adding, removing, or modifying
the buildings. The vuetify v-breadcrumbs component did not offer the
possibility to add @click events, so I created a simple one using
lists. When a user click on one of the buildings, the building (with
the floor 0 concatenated to it) gets added to the stack. When the floor
is changed, the element on the top of the stack is modified accordingly.
Finally, when the user clicks on 'main' to go back to the main view, top
entries are poped from the stack.
The navigation through the floors is implemented using the v-pagination
component. It currently shows entries staring from 1 instead of 0. This
will need to be addressed.
The code for the breadcrumb and the pagination should probably be
moved from the App component in some other component, like SCGBuildings.
Moreover, the code needs some refactoring.