Remake pagination component
Remade the pagination component using a button group toggle element.
The numeration now starts from 0 and not from 1 (this was the main
problem of the old pagination component). The bug where the floor
would not get reset to 0 when clicking on a building is also solved.
Changed the api call in api.js to only retrieve lectures of current
day.
Made a small addition to the styling of the main building so that
click effects are displayed on touch devices.